requirements.txt
argparse==1.2.1
colorama==0.3.1
dateutils==0.6.6
lxml==3.3.5
py==1.4.20
pytest==2.5.2
python-dateutil==2.2
pytz==2014.4
six==1.7.2
